For bug #217 -- admin interface for the deletion of flexbooks, their mapping to chapters and possibly the chapters as well.  Only delete the chapters if 1) the chapter is unique to the flexbook in questions, and 2) the user requests the chapters to be deleted.
Everything is done with the exception of any post-delete actions that
need to occur in helper function _do_post_delete() ... for example, the
search indices via contrib.webservice.rest.handlers.Delete ...

Also added an "Admin Home" link at the top of the admin pages while I was
at it.

Please test this stuff and let me know if there are any problems.