A number of improvements for the My Library refactor: * Upgrade from jQuery 1.2.6 to jQuery 1.3b2 * Upgrade to jQuery UI from SVN as of 20090109 * Switch from LiveQuery to jQuery 1.3b2's $.live() -- this is a big   performance boost as the former uses timers while the latter event bubbling.   (Note the "init" hack due to slightly less flexible syntax... it's an   evil, but a lesser evil.) * A rudimentary stab at preserving state of the minieditor from one page   load to the next.  Data is sent to/from the server via Django's Sessions. * Add a "sessiondata" page for quicky debugging. * A little code cleanup.
Note that jQuery 1.3 final is expected on Wednesday, Jan 14th and jQuery 1.6
UI final a week later.  On Thursday, I expect to be upgrading to jQuery 1.3
final and jQuery UI 1.6rc5.  At that time, I'll get rid of the jQuery and UI
bits in the tree.  This will probably break some CSS stuff briefly.