The following scenario wasn't working (as per previous design):
1) User searches, finds a book and adds to their minieditor.
2) User saves new book with existing chapters.
3) User goes to jQuery My Library and creates new flexbook.
4) User adds previously searched and saved book to new flexbook.

Since the chapters weren't in the list of their chapters, the add
was being ignored.  Now I know that a user should be able to have
books with chapters that aren't in their library.

Good catch Andy, thanks.  Of course, if this ends up not being the
bug you were referring to, let me know.