Bug #547 -- Abuse Prevention -- Added a "fbm_abuse_status" to the "flexbook_media" table, an enum, with four states (so far):
* ok - no abuse
* disable - image is disabled for an unspecified reason
* copyright - image is in copyright violation
* inappropriate - image content is inappropriate

I also figured out that I can use the images stored in the database
via Django, which I've added as an admin-only url.  The nice part is
that regardless of the abuse state (once Andy adds the handling), that
url will still return the original image.

I've tied this feature in with the general admin interface and provided
simple search and filtering capabilities for it.

Task remaining:
* Steve needs to produce placeholder images.
* Andy needs to wire in the new field and return Steve's images as needed.