Got the search indexing working again.  This was a little tricky and may yet have a bit to go.  I saw that src/contrib/django_restapi/responder.py was using the old ObjectPaginator and on further inspection saw that django_restapi was outdated in general, so I updated our copy (after of course checking svn logs and seeing that we had never changed it).  Then I saw that search indexing wasn't working with happy little errors like:
  2009-03-21 21:06:26,040 ERROR     contrib.webservice.rest.indexer Invalid
  parameter value(s): index.clean_and_sync; reason: 'ascii' codec can't
  encode character u'\u2019' in position 1255: ordinal not in range(128)

And u'\u2019', as it turns out, is that stupid right single quotation mark
that you get from a program like Microsoft Word... and you know how I feel
about that. So, found a happy little helper in unicode_to_ascii.py to fix
up stuff like that.  (There are potentially many bugs related to unicode
that we're going to run into due to the upgrade, but they take us towards
eventual internationalization.)

Now, what's I've done in handlers.py seems to work, but I'm marking it as
something we should look at again, if there's a better way... perhaps
when someone a little more intimate with solr (*cough*Ravi*cough*) gets
a chance.

The other two changes to handlers.py are better usage of list comprehensions.
One was making a generator that was then used in a debug statement, so
nothing meaningful was being printed.