Much of the back-end for user registration.  Besides saving additional info, it also makes a user profile and adds the user to the default user group (both of which were previously done).  I could also have these actions triggered by a signal, but this will do for now.
Sends the email as well, but I couldn't do the same domain_override
trick as before (not yet, at least).  Still need to handle a failure
to sending email as well.

But hey, if you've read this far, a handy-dandy trick in case anyone
ever wants to have a simple mail server for testing, you can use the
trick found at

http://docs.djangoproject.com/en/dev/topics/email/#testing-e-mail-sending