Add a "featured" checkbox to the Flexbook admin page.  Checking and unchecking it adds and removes the flexbook from the list of Featured flexbooks.  When adding, it populates the fields with the flexbook's state.
Here's the deal... this work is specific to Django 1.0 and will not
work with 0.96.  A little (namely set_featured_item()) will carry over,
but the FlexbookForm method of doing this is new in 1.0.  To make matters
worse, the best way to do this in 0.96 is by adding a boolean field to
the Flexbook table to manage this.  If we want this in 0.96, I'll have
to add the field, make use of it, and remove it once we upgrade to 1.0.

What I'm checking in here only adds the feature for Flexbooks, as that's
all we're using at the moment.  Adding chapters is trivial... make
chapter-specific version of FlexbookForm and add a "chapter" case to
set_featured_item().  Granted, it's a little sucky to have to have
two nearly identical forms around, but thems the breaks.