For #614 -- Make a view for mapping a perma_id with a search result. Once #1102 went down, it was a matter of adjusting the search args passed to get the desired results.  In the end, we have a url that looks like:
  http://<SERVER>/flexr/permalink/<perma_id>/

This should return at most one book or one chapter that matches the
given perma_id (or none if there no matches).  One thing that will
need to be addressed is web results, which isn't really pertinent in
this case (or in the organization serach results, for that matter).

For wrapping up this bug, we'll need to provide permalinks in the
My Library and Search Results.