UI_added 'contributor' and 'source' to attribution type input dropdown.
note: new attribution type needs to display inside of the div 'type' defined in chapter.html and flexbook.html.  currently, this is not happening.

per bug 1529, 'organization' needs to be added as well, but this requires addition to models.py and would break the app if i added it to the HTM beforehand.