UI_flexbook and chapter editor required title and description fields: 1.  adjusted styling for twirldown, divs. 2. added character counter HTML/CSS, functionality needed
VV:  there is a bug in that when you very first click 'edit', the twirldown, or the artifact, the expanded input field displays blank.  2nd clicks and beyond work ok.

VV: also, b/c of change to the HTML, the JS for chapter editor needs to be adjusted so that 'apply changes' works again

VV/NK:  added character counter to actions area in expanded view.  Functionality needs to be added here.