#1785: Editor - book titles can not be duplicated; error displayed when saved with same title https://insight.ck12.org/bugs/show_bug.cgi?id=1785 #2502 : FLexbook Editor - Ordering of book is not working if there is only one chapter
added check for duplicate flexbook titles.
a dialog is shown to the user to decide whether he wants to change the title or continue with the same title.
clicking on edit title from dialog opens the title editor with current text selected. user can start editing right away

moved ghost rows into a different container. this fixes #2502 and all conditions that may push a row to the bottom of the container.