#2708: Save of a chapter should update the appropriate flexbook accordingly on chapter save.
flexbook editor now picks up the dirty state from minieditor. if there are any chapters added to the flexbook editor in the minieditor, the flexbook editor will start as dirty.
dirty flexbooks prompt users to save the flexbook before continuing with the chapter edit.
chapter editor now receives fbid parameter when opened from flexbook editor or flexbook details.
this fbid is saved to the chapter save handler.
on chapter save, session flexbook is updated with derived chapter id.
if chapter save goes through but reference fails to update, we prompt the user to exit the chapter editor and save the flexbook again.