#2773: ERROR     flexr.middleware 'Dispatcher' object has no attribute 'func_name' using callback.__name__ instead of  callback.func_name.
this exception was getting triggerred when indexer rest api was being accessed by workbench. that that request is not an ajax request and calback.func_name is undefined for the matched url pattern.
using callback.__name__ prevents this exception. this does not affect the way breadcrumbs are rendered or managed.