#2776: On Library Delete: webservice.rest.indexer Flexbook matching query does not exist
this exception was getting triggered while deleting printjobs that belong to deleted flexbook. workbench was trying to update the solr indices for the deleted flexbook leading to this exception.

modified workbench to skip solr index update on printjob delete if respective flexbook does not exist.
also modified the debug log messege to be more obvious.