UI_PR.SR_simplified registration_added styles to hint and error messages.  styled agreement message.  added style for default title in input boxes, consistent with the metadata attribution input boxes.
note to RG/NK2:  since the class for 'inputstart' has been added to py file, it is applied to all states of the input, but should be moved to the JS so it is only in effect before user inputs their own text.  this is the same behaviour implemented on the attribution metadata inputs.  also, can we add a default title to the email and password boxes as well for consistency?