#3796: Video Shows in Preview box while using embed code in URL box from Insert Media #3792: Wrong messege is shown for inserting empty Embed Code in insert media dialog
Updated the generatePreview method to better identify URLs vs Embed Codes and switch to appropriate input type
when an embed code is inserted in URL field, it switches to Embed Code input and vice versa in case of a url placed in embed code field.
also added classnames to the radio buttons and modified the current input type selection check.