Bug 10233 - [Concept Editor] - Section format is changed after copy and paste
Heading 1 to Heading 6 are valid allowed elements. Within tinyMCE we change the headings if data paste from external source having bigger headings.
 
Using Format option text can be changed to headings and 3 sections are allowed as section 1 2 & 3. These section makes text as header h3,h4,h5 respectively. 

Due to this issue occurs that if the text copied from editor which is section 1 i.e h3 and paste it will get changed to h4.
 
For paste within tinyMCE 

h1 - is changed to - h2
h2 - is changed to - h3
h3 - is changed to - h4 (Section 1)
h4 - Nothing specific (kept as it  is) (section 2)
h5 - is changed to - h4 (section 3 )
h6 - is changed to - h4

So the solution for this can be as follows

1) We should not change the headers for sections i.e h3 to h5 it should be rosseta friendly and  should be allowed. 
2) Add header replace for h4 to h4 which is not there as of now (else Bug 9066 will occur for section 2 i.e. h4)
3) Change header replace for h5 (section 2)to keep it as h5 instead changing it to h4
4) Change h6 to h5 instead of changing it to h4 as else paste data h6 will look bold than paste of h5

So the resultant  paste within tinyMCE will be as follows 

h1 - is changed to - h2
h2 - is changed to - h3
h3 - is changed to - h3 (Section 1)
h4 - is changed to - h4 (section 2)
h5 - is changed to - h5 (section 3)
h6 - is changed to - h5