Good news: swimlanes render correctly.  Bad news: It thinks a connected graph is disconnected.
The good news: It renders the swimlanes correctly.  This is absolutely
huge news; it means that Sameer's dynamic concept map is possible.

The bad news: The autolayout algorith is reporting that bio.cxl is
disconnected, that it contains 11 (!) distinct graphs.  However,
when I find the distinct graph with the "Biology" root, it reports that
the size of that graph is the size of the dataset (WTF?), not a subset
(where are the disconnects then?).

Supposition: the getConnectedComponents() algorithm is looking for
"strongly" connected components.   I saw that term in the literature; I
will have to research what it means.  *Sigh*