Adjustments, curved lines, and all the other happiness.
If you want something *really* fun, load this up in a very big monitor,
set the zoom factor to 50%, then run 'Start Twiddler.'  You'll be able to
watch it animate as it tries to figure out the best visual presentation
(lines from unrelated topics not crossing, things like that).
Eventually it reaches an equilibrium.  I haven't figured out yet how to
detect equilibrium, so it jitters for a while until it reaches an
arbitrary end.

The animation is *not* for production.  It's just something I needed to
be able to visualize the layout algorithm when it was broken.