These additions allow flxweb, which is a pure python program, to be run in a VirtualEnv (http://www.virtualenv.org/en/latest/index.html) python isolated environment.  Other than the Python version installed on your OS and the Python Standard Library, every other requirement will be applicable only to the current local environment.
The file requirements.txt isolates further the *exact* environment
in which FlxWeb runs: it specifies version numbers of required files,
allowing you to verify a "last known good" version number for our work.
It also allows us to nail down version numbers against work-arounds.
It's embarassing when a future version breaks an existing implementation
because the existing implementation has work-arounds for known bugs in
the version against which it was developed.

The .gitignore file allows git-svn users to work in this environment
without accidentally checking in .pyc, .pyo, various backups, logs and
other generated files that shouldn't be part of the check-in.