Integration proceeds apace.  We're now pulling the data correctly, in the correct format, with correct DAGs.  However, the pan-and-scan stuff is limiting the page draw size, and I'm not sure why.  There's a lot of fiddly bits here that I *think* I can cut down-- mostly I discovered that the canvas/tdiv interface was one level too abstract; I can cut out the tdiv and put concepts directly onto the parent div as sibling of the canvas, and it seems to work fine in most browsers as long as the z-index of the canvas is always less than that of any concept DOM objects.
Added some caching to the concept handler (4 minutes... enough to make
quick reloads less annoying in a build/debug cycle), but that can be
taken out.  *Very* annoyed that cache busting isn't working with
require, and still puzzled as to why require sometimes looks for jquery
and backbone in the wrong path, but only sometimes.  I *am* loading
require.js first, after all, and using the examples set by the rest of
the JS team.