I'm having a real headache with the viewport size issue.  I don't understand (yet) why it's height is solid, but its width keeps changing; it ought not to change, since it's locked to the grid Sling picked.
But this does manage to draw things correctly.  Problem is, it's *slow*.
Like 3 second pause slow on a fast machine.  That's mostly because the
layout-and-render pass happens twice-- once when the data is loaded, and
a second time when the damn viewport finally figures out how big it is.

I'll research, but for now, this seems to be a solution that works.