Bug 12195: call isotope.reLayout() whenever hiding the concept display to force a reflow.
Also, wrap the reflow command in an 'if' statement to distinguish
between a 'hide concepts' event and a 'show different concepts' event.
This prevents spamming the jQuery event dispatch hub, which could have
performance-related consequences.

I don't know that we need isotope here at all, since we never reflow the
browsecards, but this was the simplest fix.