Added a filter to modality.py so that if the modalityCount data type isn't a dictionary it doesn't crash.
Added a filter to browse.py so that empty modality lists don't crash the
render.

Added a filter to browse.py so that modality counts of zero don't get
forwarded to the render.

Added a filter to browse.py to filter out 'concept' from the modalities
handlers, since 'lesson' duplicates it and incorrectly inflates the
count for modality group 'Text'.

Note: All of these filters are necessary only because the datasets from
Flx are so poor.  The top issue is a transmission protocol schema failure.
The second and third are data discipline failures.  The fourth is
a database schema failure, probably caused by an evolutionary step.

The modality group 'Teaching Resources' had 'rubric' misspelled.  This
caused incorrect counts in Nak's original linear search; when I switched
to a lookup table it caused a crash (CE/CO).  Fixed.