(1) Provide an empty "browse_term" context and catch the API failures, such that when a given browse term has no BrowseTermDescendents the user is presented with the "Help us add concepts for..." message.
(2) Add the "Help us add concepts for..." message.  Provide it when the
"browse_term" is empty.  This can happen when either there's an API
failure at the Flx end (the subject doesn't exist in that portion of the
database-- this, by the way, is an SQL level fail) or the subject does
exist but no descendent browse terms have been provided.