Fix a jshint complaint.
This, by the way, illustrates the problem with doing the checks using a
client-side hook.  I was able to fix the file and *not* add it to the
index of files to be committed server-side, and the client-side hook let
it pass because the file client-side look fine.  I had to make a second
commit to include the fixed version.

This problem is accurately covered here:
http://tech.yipit.com/2011/11/16/183772396/

Please, pretty please, implement the server-side hooks?